Move DTOs to bee-message
