+++ bee-api/bee-rest-api/src/endpoints/routes/api/v2/submit_message.rs
@@ -9,10 +9,10 @@ use crate::{
         rejection::CustomRejection,
         storage::StorageBackend,
     },
-    types::{dtos::PayloadDto, responses::SubmitMessageResponse},
+    types::{responses::SubmitMessageResponse},
 };
 
-use bee_message::{parent::Parents, payload::Payload, Message, MessageBuilder, MessageId};
+use bee_message::{parent::Parents, payload::{Payload, dto::PayloadDto}, Message, MessageBuilder, MessageId};
 use bee_pow::providers::{miner::MinerBuilder, NonceProviderBuilder};
 use bee_protocol::{
     workers::{config::ProtocolConfig, MessageSubmitterError, MessageSubmitterWorkerEvent},
