+++ bee-message/src/address/mod.rs
@@ -148,3 +148,108 @@ impl Address {
         Ok(())
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize, Serializer};
+    use serde_json::Value;
+
+    use super::*;
+    pub use super::{alias::dto::AliasAddressDto, ed25519::dto::Ed25519AddressDto, nft::dto::NftAddressDto};
+    use crate::error::dto::DtoError;
+
+    /// Describes all the different address types.
+    #[derive(Clone, Debug)]
+    pub enum AddressDto {
+        /// An Ed25519 address.
+        Ed25519(Ed25519AddressDto),
+        /// An alias address.
+        Alias(AliasAddressDto),
+        /// A NFT address.
+        Nft(NftAddressDto),
+    }
+
+    impl From<&Address> for AddressDto {
+        fn from(value: &Address) -> Self {
+            match value {
+                Address::Ed25519(a) => AddressDto::Ed25519(a.into()),
+                Address::Alias(a) => AddressDto::Alias(a.into()),
+                Address::Nft(a) => AddressDto::Nft(a.into()),
+            }
+        }
+    }
+
+    impl TryFrom<&AddressDto> for Address {
+        type Error = DtoError;
+
+        fn try_from(value: &AddressDto) -> Result<Self, Self::Error> {
+            match value {
+                AddressDto::Ed25519(a) => Ok(Address::Ed25519(a.try_into()?)),
+                AddressDto::Alias(a) => Ok(Address::Alias(a.try_into()?)),
+                AddressDto::Nft(a) => Ok(Address::Nft(a.try_into()?)),
+            }
+        }
+    }
+
+    impl<'de> Deserialize<'de> for AddressDto {
+        fn deserialize<D: serde::Deserializer<'de>>(d: D) -> Result<Self, D::Error> {
+            let value = Value::deserialize(d)?;
+            Ok(
+                match value
+                    .get("type")
+                    .and_then(Value::as_u64)
+                    .ok_or_else(|| serde::de::Error::custom("invalid address type"))? as u8
+                {
+                    Ed25519Address::KIND => {
+                        AddressDto::Ed25519(Ed25519AddressDto::deserialize(value).map_err(|e| {
+                            serde::de::Error::custom(format!("cannot deserialize ed25519 address: {}", e))
+                        })?)
+                    }
+                    AliasAddress::KIND => {
+                        AddressDto::Alias(AliasAddressDto::deserialize(value).map_err(|e| {
+                            serde::de::Error::custom(format!("cannot deserialize alias address: {}", e))
+                        })?)
+                    }
+                    NftAddress::KIND => AddressDto::Nft(
+                        NftAddressDto::deserialize(value)
+                            .map_err(|e| serde::de::Error::custom(format!("cannot deserialize NFT address: {}", e)))?,
+                    ),
+                    _ => return Err(serde::de::Error::custom("invalid address type")),
+                },
+            )
+        }
+    }
+
+    impl Serialize for AddressDto {
+        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+        where
+            S: Serializer,
+        {
+            #[derive(Serialize)]
+            #[serde(untagged)]
+            enum AddressDto_<'a> {
+                T1(&'a Ed25519AddressDto),
+                T2(&'a AliasAddressDto),
+                T3(&'a NftAddressDto),
+            }
+            #[derive(Serialize)]
+            struct TypedAddress<'a> {
+                #[serde(flatten)]
+                address: AddressDto_<'a>,
+            }
+            let address = match self {
+                AddressDto::Ed25519(o) => TypedAddress {
+                    address: AddressDto_::T1(o),
+                },
+                AddressDto::Alias(o) => TypedAddress {
+                    address: AddressDto_::T2(o),
+                },
+                AddressDto::Nft(o) => TypedAddress {
+                    address: AddressDto_::T3(o),
+                },
+            };
+            address.serialize(serializer)
+        }
+    }
+}
