+++ bee-message/src/address/nft.rs
@@ -54,3 +54,41 @@ impl core::fmt::Debug for NftAddress {
         write!(f, "NftAddress({})", self)
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::error::dto::DtoError;
+
+    /// Describes an NFT address.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct NftAddressDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        #[serde(rename = "nftId")]
+        pub nft_id: String,
+    }
+
+    impl From<&NftAddress> for NftAddressDto {
+        fn from(value: &NftAddress) -> Self {
+            Self {
+                kind: NftAddress::KIND,
+                nft_id: value.to_string(),
+            }
+        }
+    }
+
+    impl TryFrom<&NftAddressDto> for NftAddress {
+        type Error = DtoError;
+
+        fn try_from(value: &NftAddressDto) -> Result<Self, Self::Error> {
+            value
+                .nft_id
+                .parse::<NftAddress>()
+                .map_err(|_| DtoError::InvalidField("NFT address"))
+        }
+    }
+}
