+++ bee-message/src/output/alias.rs
@@ -535,3 +535,101 @@ fn verify_unlock_conditions(unlock_conditions: &UnlockConditions, alias_id: &Ali
 
     verify_allowed_unlock_conditions(unlock_conditions, AliasOutput::ALLOWED_UNLOCK_CONDITIONS)
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::{
+        error::dto::DtoError,
+        output::{
+            alias_id::dto::AliasIdDto, feature_block::dto::FeatureBlockDto, native_token::dto::NativeTokenDto,
+            unlock_condition::dto::UnlockConditionDto,
+        },
+    };
+
+    /// Describes an alias account in the ledger that can be controlled by the state and governance controllers.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct AliasOutputDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        // Amount of IOTA tokens held by the output.
+        pub amount: String,
+        // Native tokens held by the output.
+        #[serde(rename = "nativeTokens")]
+        pub native_tokens: Vec<NativeTokenDto>,
+        // Unique identifier of the alias.
+        #[serde(rename = "aliasId")]
+        pub alias_id: AliasIdDto,
+        // A counter that must increase by 1 every time the alias is state transitioned.
+        #[serde(rename = "stateIndex")]
+        pub state_index: u32,
+        // Metadata that can only be changed by the state controller.
+        #[serde(rename = "stateMetadata")]
+        pub state_metadata: String,
+        // A counter that denotes the number of foundries created by this alias account.
+        #[serde(rename = "foundryCounter")]
+        pub foundry_counter: u32,
+        //
+        #[serde(rename = "unlockConditions")]
+        pub unlock_conditions: Vec<UnlockConditionDto>,
+        //
+        #[serde(rename = "featureBlocks")]
+        pub feature_blocks: Vec<FeatureBlockDto>,
+        //
+        #[serde(rename = "immutableFeatureBlocks")]
+        pub immutable_feature_blocks: Vec<FeatureBlockDto>,
+    }
+
+    impl From<&AliasOutput> for AliasOutputDto {
+        fn from(value: &AliasOutput) -> Self {
+            Self {
+                kind: AliasOutput::KIND,
+                amount: value.amount().to_string(),
+                native_tokens: value.native_tokens().iter().map(Into::into).collect::<_>(),
+                alias_id: AliasIdDto(value.alias_id().to_string()),
+                state_index: value.state_index(),
+                state_metadata: prefix_hex::encode(value.state_metadata()),
+                foundry_counter: value.foundry_counter(),
+                unlock_conditions: value.unlock_conditions().iter().map(Into::into).collect::<_>(),
+                feature_blocks: value.feature_blocks().iter().map(Into::into).collect::<_>(),
+                immutable_feature_blocks: value.immutable_feature_blocks().iter().map(Into::into).collect::<_>(),
+            }
+        }
+    }
+
+    impl TryFrom<&AliasOutputDto> for AliasOutput {
+        type Error = DtoError;
+
+        fn try_from(value: &AliasOutputDto) -> Result<Self, Self::Error> {
+            let mut builder = AliasOutputBuilder::new(
+                value
+                    .amount
+                    .parse::<u64>()
+                    .map_err(|_| DtoError::InvalidField("amount"))?,
+                (&value.alias_id).try_into()?,
+            )?;
+            builder = builder.with_state_index(value.state_index);
+            builder = builder.with_state_metadata(
+                prefix_hex::decode(&value.state_metadata).map_err(|_| DtoError::InvalidField("state_metadata"))?,
+            );
+            builder = builder.with_foundry_counter(value.foundry_counter);
+
+            for t in &value.native_tokens {
+                builder = builder.add_native_token(t.try_into()?);
+            }
+            for b in &value.unlock_conditions {
+                builder = builder.add_unlock_condition(b.try_into()?);
+            }
+            for b in &value.feature_blocks {
+                builder = builder.add_feature_block(b.try_into()?);
+            }
+            for b in &value.immutable_feature_blocks {
+                builder = builder.add_immutable_feature_block(b.try_into()?);
+            }
+            Ok(builder.finish()?)
+        }
+    }
+}
