+++ bee-message/src/output/basic.rs
@@ -208,3 +208,69 @@ fn verify_feature_blocks<const VERIFY: bool>(blocks: &FeatureBlocks) -> Result<(
         Ok(())
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::{
+        error::dto::DtoError,
+        output::{
+            feature_block::dto::FeatureBlockDto, native_token::dto::NativeTokenDto,
+            unlock_condition::dto::UnlockConditionDto,
+        },
+    };
+
+    /// Describes a basic output.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct BasicOutputDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        // Amount of IOTA tokens held by the output.
+        pub amount: String,
+        // Native tokens held by the output.
+        #[serde(rename = "nativeTokens")]
+        pub native_tokens: Vec<NativeTokenDto>,
+        #[serde(rename = "unlockConditions")]
+        pub unlock_conditions: Vec<UnlockConditionDto>,
+        #[serde(rename = "featureBlocks")]
+        pub feature_blocks: Vec<FeatureBlockDto>,
+    }
+
+    impl From<&BasicOutput> for BasicOutputDto {
+        fn from(value: &BasicOutput) -> Self {
+            Self {
+                kind: BasicOutput::KIND,
+                amount: value.amount().to_string(),
+                native_tokens: value.native_tokens().iter().map(Into::into).collect::<_>(),
+                unlock_conditions: value.unlock_conditions().iter().map(Into::into).collect::<_>(),
+                feature_blocks: value.feature_blocks().iter().map(Into::into).collect::<_>(),
+            }
+        }
+    }
+
+    impl TryFrom<&BasicOutputDto> for BasicOutput {
+        type Error = DtoError;
+
+        fn try_from(value: &BasicOutputDto) -> Result<Self, Self::Error> {
+            let mut builder = BasicOutputBuilder::new(
+                value
+                    .amount
+                    .parse::<u64>()
+                    .map_err(|_| DtoError::InvalidField("amount"))?,
+            )?;
+            for t in &value.native_tokens {
+                builder = builder.add_native_token(t.try_into()?);
+            }
+            for b in &value.unlock_conditions {
+                builder = builder.add_unlock_condition(b.try_into()?);
+            }
+            for b in &value.feature_blocks {
+                builder = builder.add_feature_block(b.try_into()?);
+            }
+            Ok(builder.finish()?)
+        }
+    }
+}
