+++ bee-message/src/output/feature_block/issuer.rs
@@ -26,3 +26,18 @@ impl IssuerFeatureBlock {
         &self.0
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use crate::address::dto::AddressDto;
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct IssuerFeatureBlockDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub address: AddressDto,
+    }
+}
