+++ bee-message/src/output/feature_block/metadata.rs
@@ -61,3 +61,16 @@ impl core::fmt::Debug for MetadataFeatureBlock {
         write!(f, "MetadataFeatureBlock({})", self)
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct MetadataFeatureBlockDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub data: String,
+    }
+}
