+++ bee-message/src/output/feature_block/tag.rs
@@ -61,3 +61,16 @@ impl core::fmt::Debug for TagFeatureBlock {
         write!(f, "TagFeatureBlock({})", self)
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct TagFeatureBlockDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub tag: String,
+    }
+}
