+++ bee-message/src/output/nft_id.rs
@@ -17,6 +17,36 @@ impl From<OutputId> for NftId {
 impl NftId {
     ///
     pub fn or_from_output_id(self, output_id: OutputId) -> Self {
-        if self.is_null() { Self::from(output_id) } else { self }
+        if self.is_null() {
+            Self::from(output_id)
+        } else {
+            self
+        }
+    }
+}
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::error::dto::DtoError;
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct NftIdDto(pub String);
+
+    impl From<&NftId> for NftIdDto {
+        fn from(value: &NftId) -> Self {
+            Self(value.to_string())
+        }
+    }
+
+    impl TryFrom<&NftIdDto> for NftId {
+        type Error = DtoError;
+
+        fn try_from(value: &NftIdDto) -> Result<Self, Self::Error> {
+            value.0.parse::<NftId>().map_err(|_| DtoError::InvalidField("NFT id"))
+        }
     }
 }
