+++ bee-message/src/output/treasury.rs
@@ -38,3 +38,42 @@ impl TreasuryOutput {
         self.amount.get()
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::error::dto::DtoError;
+
+    /// Describes a treasury output.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct TreasuryOutputDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub amount: String,
+    }
+
+    impl From<&TreasuryOutput> for TreasuryOutputDto {
+        fn from(value: &TreasuryOutput) -> Self {
+            Self {
+                kind: TreasuryOutput::KIND,
+                amount: value.amount().to_string(),
+            }
+        }
+    }
+
+    impl TryFrom<&TreasuryOutputDto> for TreasuryOutput {
+        type Error = DtoError;
+
+        fn try_from(value: &TreasuryOutputDto) -> Result<Self, Self::Error> {
+            Ok(Self::new(
+                value
+                    .amount
+                    .parse::<u64>()
+                    .map_err(|_| DtoError::InvalidField("amount"))?,
+            )?)
+        }
+    }
+}
