+++ bee-message/src/output/unlock_condition/address.rs
@@ -26,3 +26,18 @@ impl AddressUnlockCondition {
         &self.0
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use crate::address::dto::AddressDto;
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct AddressUnlockConditionDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub address: AddressDto,
+    }
+}
