+++ bee-message/src/output/unlock_condition/expiration.rs
@@ -101,3 +101,23 @@ fn verify_milestone_index_timestamp(milestone_index: MilestoneIndex, timestamp: 
         Ok(())
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use crate::{address::dto::AddressDto, milestone::MilestoneIndex};
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct ExpirationUnlockConditionDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        #[serde(rename = "returnAddress")]
+        pub return_address: AddressDto,
+        #[serde(rename = "milestoneIndex")]
+        pub milestone_index: MilestoneIndex,
+        #[serde(rename = "unixTime")]
+        pub timestamp: u32,
+    }
+}
