+++ bee-message/src/output/unlock_condition/governor_address.rs
@@ -27,3 +27,18 @@ impl GovernorAddressUnlockCondition {
         &self.0
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use crate::address::dto::AddressDto;
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct GovernorAddressUnlockConditionDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        pub address: AddressDto,
+    }
+}
