+++ bee-message/src/output/unlock_condition/timelock.rs
@@ -84,3 +84,21 @@ fn verify_milestone_index_timestamp(milestone_index: MilestoneIndex, timestamp: 
         Ok(())
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use crate::milestone::MilestoneIndex;
+
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct TimelockUnlockConditionDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        #[serde(rename = "milestoneIndex")]
+        pub milestone_index: MilestoneIndex,
+        #[serde(rename = "unixTime")]
+        pub timestamp: u32,
+    }
+}
