+++ bee-message/src/payload/milestone/mod.rs
@@ -197,3 +197,102 @@ fn verify_essence_signatures(essence: &MilestoneEssence, signatures: &[Signature
         Ok(())
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+    use crate::{
+        error::dto::DtoError, milestone::MilestoneIndex, parent::Parents, payload::dto::PayloadDto, MessageId,
+    };
+
+    /// The payload type to define a milestone.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct MilestonePayloadDto {
+        #[serde(rename = "type")]
+        pub kind: u32,
+        pub index: u32,
+        pub timestamp: u64,
+        #[serde(rename = "parentMessageIds")]
+        pub parents: Vec<String>,
+        #[serde(rename = "inclusionMerkleProof")]
+        pub inclusion_merkle_proof: String,
+        #[serde(rename = "nextPoWScore")]
+        pub next_pow_score: u32,
+        #[serde(rename = "nextPoWScoreMilestoneIndex")]
+        pub next_pow_score_milestone_index: u32,
+        #[serde(rename = "publicKeys")]
+        pub public_keys: Vec<String>,
+        #[serde(skip_serializing_if = "Option::is_none")]
+        pub receipt: Option<PayloadDto>,
+        pub signatures: Vec<String>,
+    }
+
+    impl From<&MilestonePayload> for MilestonePayloadDto {
+        fn from(value: &MilestonePayload) -> Self {
+            MilestonePayloadDto {
+                kind: MilestonePayload::KIND,
+                index: *value.essence().index(),
+                timestamp: value.essence().timestamp(),
+                parents: value.essence().parents().iter().map(|p| p.to_string()).collect(),
+                inclusion_merkle_proof: prefix_hex::encode(value.essence().merkle_proof()),
+                next_pow_score: value.essence().next_pow_score(),
+                next_pow_score_milestone_index: value.essence().next_pow_score_milestone_index(),
+                public_keys: value.essence().public_keys().iter().map(prefix_hex::encode).collect(),
+                receipt: value.essence().receipt().map(Into::into),
+                signatures: value.signatures().map(prefix_hex::encode).collect(),
+            }
+        }
+    }
+
+    impl TryFrom<&MilestonePayloadDto> for MilestonePayload {
+        type Error = DtoError;
+
+        fn try_from(value: &MilestonePayloadDto) -> Result<Self, Self::Error> {
+            let essence = {
+                let index = value.index;
+                let timestamp = value.timestamp;
+                let mut parent_ids = Vec::new();
+                for msg_id in &value.parents {
+                    parent_ids.push(
+                        msg_id
+                            .parse::<MessageId>()
+                            .map_err(|_| DtoError::InvalidField("parentMessageIds"))?,
+                    );
+                }
+                let merkle_proof = prefix_hex::decode(&value.inclusion_merkle_proof)
+                    .map_err(|_| DtoError::InvalidField("inclusionMerkleProof"))?;
+                let next_pow_score = value.next_pow_score;
+                let next_pow_score_milestone_index = value.next_pow_score_milestone_index;
+                let mut public_keys = Vec::new();
+                for v in &value.public_keys {
+                    public_keys.push(prefix_hex::decode(v).map_err(|_| DtoError::InvalidField("publicKeys"))?);
+                }
+                let receipt = if let Some(receipt) = value.receipt.as_ref() {
+                    Some(receipt.try_into()?)
+                } else {
+                    None
+                };
+                MilestoneEssence::new(
+                    MilestoneIndex(index),
+                    timestamp,
+                    Parents::new(parent_ids)?,
+                    merkle_proof,
+                    next_pow_score,
+                    next_pow_score_milestone_index,
+                    public_keys,
+                    receipt,
+                )?
+            };
+            let mut signatures = Vec::new();
+            for v in &value.signatures {
+                let sig: Vec<u8> = prefix_hex::decode(v).map_err(|_| DtoError::InvalidField("signatures"))?;
+                signatures.push(sig.try_into().map_err(|_| DtoError::InvalidField("signatures"))?)
+            }
+
+            Ok(MilestonePayload::new(essence, signatures)?)
+        }
+    }
+}
