+++ bee-message/src/unlock_block/reference.rs
@@ -33,3 +33,19 @@ impl ReferenceUnlockBlock {
         self.0.get()
     }
 }
+
+#[cfg(feature = "dto")]
+#[allow(missing_docs)]
+pub mod dto {
+    use serde::{Deserialize, Serialize};
+
+    /// References a previous unlock block in order to substitute the duplication of the same unlock block data for inputs
+    /// which unlock through the same data.
+    #[derive(Clone, Debug, Serialize, Deserialize)]
+    pub struct ReferenceUnlockBlockDto {
+        #[serde(rename = "type")]
+        pub kind: u8,
+        #[serde(rename = "reference")]
+        pub index: u16,
+    }
+}
